Moved the download and delete buttons to RecordingControls
They now display nicely underneath the recording.
